rename: assets 폴더 내부 구조 변경
