+++ packages/ui/src/components/shared/NavItem.tsx
@@ -7,7 +7,7 @@ import Link from "next/link";
 import { useSelectedLayoutSegments } from "next/navigation";
 import { useState } from "react";
 
-import arrowImageUrl from "../../assets/arrow.svg";
+import arrowImageUrl from "../../assets/images/arrow.svg";
 
 /**
  * @description 내비게이션 바에서 사용하는 내비게이션 아이템 컴포넌트입니다.
